refactor/#31: DailyRankInfo long으로 변경

+++ src/main/java/com/project/dto/DailyRankInfo.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class DailyRankInfo {
     private String name;
-    private int solved;
-    private int score;
+    private long solved;
+    private long score;
     private int rank;
 }
